added :size key to output, EDN data now written with pprint
The `:size` key records the width and height of the 2d output vector.

Instead of `spit`, I am now using `io/writer` directly with `pprint` to write
the EDN data in a more nicely formatted way that prints the vector rows in one
line each. This lets me visually scan the data in 2d similar to looking at an
image.